requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
pydantic==2.9.2
langchain==0.3.1
langchain-openai==0.2.2
langchain-google-genai==2.0.0
sentence-transformers==3.1.1
httpx==0.27.2
python-dotenv==1.0.1
tenacity>=8.1.0
streamlit==1.39.0
numpy==1.26.4
nest_asyncio==1.6.0